Fuentes v0.0.2 Fuentes.Account

The Account module represents accounts in the system which are of asset, liability, or equity types, in accordance with the “accounting equation”.

Each account must be set to one of the following types:

TYPE | NORMAL BALANCE | DESCRIPTION

asset	Debit	Resources owned by the Business Entity

liability	Credit	Debts owed to outsiders

equity	Credit	Owners rights to the Assets

Each account can also be marked as a “Contra Account”. A contra account will have it’s normal balance swapped. For example, to remove equity, a “Drawing” account may be created as a contra equity account as follows:

account = %Fuentes.Account{name: "Drawing", type: "asset", contra: true}

At all times the balance of all accounts should conform to the “accounting equation”

Assets = Liabilities + Owner’s Equity

Each account type acts as it’s own ledger.

For more details see:

Wikipedia - Accounting Equation Wikipedia - Debits, Credits, and Contra Accounts

Types

t :: %Fuentes.Account{__meta__: term, amounts: [Fuentes.Amount], balance: term, contra: Boolean.t, id: term, inserted_at: term, name: String.t, type: String.t, updated_at: term}

Functions

balance(repo \\ Config.repo(), account_or_account_list, dates \\ nil)

Specs

balance(Ecto.Repo.t, [Fuentes.Account.t], Ecto.Date.t) :: Decimal.t

Computes the account balance for a list of Fuentes.Account in a given Ecto.Repo inclusive of all entries. This function is intended to be used with a list of Fuentes.Accounts of the same type. Returns Decimal type.

changeset(model, params \\ %{})

Creates a changeset requiring a :name and :type

trial_balance(repo \\ Config.repo_from_config())

Computes the trial balance for all accounts in the provided Ecto.Repo. Returns Decimal type.