Fuentes v0.0.1 Fuentes.Amount

An amount must be of either the debit or credit type to be saved to the database.

Functions

changeset(model, params \\ %{})

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

dated(query, map)

for_account(query, account)

for_entry(query, entry)

sum_type(query, type)