Fuentes v0.0.2 Fuentes.Entry

Entries are the recording of account debits and credits and can be considered as consituting a traditional accounting Journal.

Types

t :: %Fuentes.Entry{__meta__: term, amounts: term, date: Ecto.Date.t, description: String.t, id: term, inserted_at: term, updated_at: term}

Functions

balanced?(repo \\ Config.repo(), entry)

Specs

balanced?(Ecto.Repo.t, Fuentes.Entry.t) :: Boolean.t

Accepts an Fuentes.Entry and Ecto.Repo and returns true/false based on whether the associated amounts for that entry sum to zero.

changeset(model, params \\ %{})

Creates a changeset for Fuentes.Entry, validating a required :description and :date, casting an provided “debit” and “credit” Fuentes.Amounts, and validating that those amounts balance.

validate_debits_and_credits_balance(changeset)

Accepts and returns a changeset, adding an error if “credit” and “debit” amounts are not equivalent