Fuentes v0.0.1 Mix.Tasks.Fuentes.Install

Generates an Ecto model.

mix fuentes.install

or

mix fuentes.install Admin admins

The first argument is the module name for the user and the second argument is the plural form of the module name. If omitted, it uses default module name as User and its plural form users.

The generated resource will contain:

a model in web/models

a controller, a view, and templates for User registration

a controller, a view, and templates for Session management

a controller, a view, and templates for Password reset

a migration file for the repository

Functions

run(args)

Callback implementation for c:Mix.Task.run/1.