Fuentes v0.0.2 Fuentes.Amount

An Amount represents the individual debit or credit for a given account and is part of a balanced entry.

Types

t :: %Fuentes.Amount{__meta__: term, account: term, account_id: Integer.t, amount: Decimal.t, entry: term, entry_id: Integer.t, id: term, inserted_at: term, type: String.t, updated_at: term}

Functions

changeset(model, params \\ %{})

Creates an amount changeset associated with a Fuentes.Entry and Fuentes.Account. A type (“credit” or “debit”), as well as, an amount greater than 0 must be specified.