Fuentes v0.0.1 Fuentes.Entry

Entries are the recording of account debits and credits and can be considered as consituting a traditional accounting Journal.

Functions

balanced?(entry, repo)

changeset(model, params \\ %{})

Creates a changeset based on the model and params.

If no params are provided, an invalid changeset is returned with no validation performed.

validate_debits_and_credits_balance(changeset)